% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusp_fit.R
\name{dcusp}
\alias{dcusp}
\title{Dipole-dipole Cusp Line Shape}
\usage{
dcusp(x, width, xoff)
}
\arguments{
\item{x}{a numeric. The position at which the distribution is calculated.}

\item{width}{a numeric. The width of the line shape. The width parameter is
related to the full-width at half-maximum (FWHM) by width = FWHM / 3.09.}

\item{xoff}{a numeric. The offset of the distribution along the x-axis.}
}
\description{
Returns a cusp-like line shape
}
\details{
\code{dcusp} provides the cusp-like line shape used in
\href{http://journals.aps.org/pra/abstract/10.1103/PhysRevA.93.042505}{Dipole-dipole resonance line shapes in a cold Rydberg gas} by Richards and Jones.
}

