% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorentzian_fit.R
\name{fitCauchy}
\alias{fitCauchy}
\alias{fitCauchy2}
\title{Lorentzian/Cauchy Fit}
\usage{
fitCauchy(x, y, center, hwhm, height, yoff)

fitCauchy2(x, y, center, hwhm, height)
}
\arguments{
\item{x}{A numeric vector. The x-axis of the data to be fit.}

\item{y}{A numeric vector. The y-axis of the data to be fit.}

\item{center}{A numeric. A starting guess for the center of the Lorentzian.}

\item{hwhm}{A numeric. A starting guess for the half-width at half-max of
the Lorentzian.}

\item{height}{A numeric. A starting guess for the ampltidue scaling of the
Lorentzian.}

\item{yoff}{A numeric. A starting guess for the offset from the y-axis.}
}
\description{
\code{fitCauchy} calculates the Lorentzian/Cauchy fit
}
\details{
This function calculates the Lorentzian/Cauchy fit of the provided data. It
uses \code{\link[stats]{dcauchy}} to calculate the Lorentzian/Cauchy
profile and optimizes it. This version contains a y-offset optimization.
}
\section{Functions}{
\itemize{
\item \code{fitCauchy2}: A Lorentzian/Cauchy fit without the y-offset.
}}

