% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusp_fit.R
\name{fitCusp}
\alias{fitCusp}
\alias{fitCusp2}
\alias{fitCusp3}
\title{Cusp Line Shape Fit}
\usage{
fitCusp(x, y, height, width, xoff, yoff)

fitCusp2(x, y, height, width, xoff, yoff)

fitCusp3(x, y, height, width, xoff)
}
\arguments{
\item{x}{a numeric vector. The x coordinates to be fit.}

\item{y}{a numeric vector. The y coordinates to be fit.}

\item{height}{a numeric. An initial guess to the height of the distribution.}

\item{width}{a numeric. An initial guess to the width of the distribution.}

\item{xoff}{a numeric. An initial guess to the x-offset of the
distribution.}

\item{yoff}{a numeric. An initial guess to the y-offset of the
distribution.}
}
\description{
Performs a fit to the cusp line shape \code{\link{dcusp}}
}
\details{
These functions perform fits of provided data to the cusp line shape
\code{\link{dcusp}} using \code{\link[stats]{optim}}. The base version
of \code{fitCusp} using a max iteration number of 300 in \code{optim}.
\code{fitCusp2} uses a max iteration number of 1000. \code{fitCusp3}
also uses a max iteration number of 1000, but removes the y-offset.
}
