% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_fit.R
\name{fitG}
\alias{fitG}
\alias{fitG2}
\title{Gaussian Fit}
\usage{
fitG(x, y, mu, sig, scale, yoff)

fitG2(x, y, mu, sig, scale)
}
\arguments{
\item{x}{A numeric vector. The x-axis of the data to be fit.}

\item{y}{A numeric vector. The y-axis of the data to be fit.}

\item{mu}{A numeric. A starting guess for the mean of the Gaussian.}

\item{sig}{A numeric. A starting guess for the standard deviation of the
Gaussian.}

\item{scale}{A numeric. A starting guess for the ampltidue scaling of the
Gaussian.}

\item{yoff}{A numeric. A starting guess for the offset from the y-axis.}
}
\description{
\code{fitG} calculates the Gaussian fit
}
\details{
This function calculates the Gaussian fit of the provided data. It uses
\code{\link[stats]{dnorm}} calculate the Gaussian profile and optimizes
it. This version contains a y-offset optimization.
}
\section{Functions}{
\itemize{
\item \code{fitG2}: A Gaussian fit without the y-offset.
}}

