% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/width_measurement.R
\name{FWHM}
\alias{FWHM}
\alias{HWHMleft}
\alias{HWHMright}
\title{Width Measurement (FWHM and HWHM)}
\usage{
FWHM(spectrum)

HWHMleft(spectrum)

HWHMright(spectrum)
}
\arguments{
\item{spectrum}{a numeric matrix. The x and y coordinates of the function
whose width is to be measured.}
}
\description{
Returns the width of a given spectrum.
}
\details{
These functions provide the full-width at half-maximum or the half-width
at half-maximum depending on the function. Given a two column matrix
describing the x and y positions of a spectrum, the function finds the
location of the maximum and finds the spot just below the half-max point
on either side of the peak. The function \code{\link[stats]{approx}} is
used to determine the actual location of the half-max point on either
side of the maximum. The separation of these two points is then computed
provided the full maximum. For the half-width functions, onlye the right
side is computed in \code{HWHMright} and only the left side is computed
in \code{HWHMleft}.
}

