% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\docType{data}
\name{report_theme}
\alias{report_theme}
\title{Report Theme}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 44.}
\usage{
report_theme
}
\description{
Theme
}
\details{
Theme Descript
}
\keyword{datasets}

