% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/width_measurement.R
\name{WidthMeasure}
\alias{WidthMeasure}
\title{Width Measurement}
\usage{
WidthMeasure(spectrum)
}
\arguments{
\item{spectrum}{a numeric matrix. The x and y coordinates of the function
whose width is to be measured.}
}
\description{
Returns the full-width at half-maximum
}
\details{
This function returns the full-width at half-maximum of a given spectrum.
It initially calls \code{\link{FWHM}}. If this function provides a NA width,
then \code{\link{HWHMleft}} is called and multiplied by 2. If that also
returns a NA width, \code{\link{HWHMright}} is called and multiplied by 2.
}

